TOP stock_picker assignment
This method takes an array of numbers that represent prices and returns
an array of two numbers, the first being the best day to to buy and the
second being the best day to sell.

This method runs through each element in the array, finds its largest
potential amount that it can earn (negative amounts round to 0) and
puts the indexes of the buy/sell days into the final array.